# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SIM Service API",
    "description": "API untuk Sistem Informasi Manajemen dengan fitur CRUD User, KTP (Kartu Tanda Penduduk), dan SIM (Surat Izin Mengemudi). Mendukung autentikasi JWT, pagination, search, dan sorting.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8026",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "description": "ID unik user (auto-generated dengan nanoid)",
            "example": "V1StGXR8_Z5jdHi6B-myT"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email user",
            "example": "john.doe@example.com"
          },
          "full_name": {
            "type": "string",
            "description": "Nama lengkap user",
            "example": "John Doe"
          },
          "is_active": {
            "type": "boolean",
            "description": "Status aktif user",
            "example": true
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "status_code": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "Pesan response",
            "example": "Operasi berhasil"
          },
          "data": {
            "type": "object",
            "description": "Data response"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "status_code": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 400
          },
          "errors": {
            "oneOf": [
              {
                "type": "string",
                "description": "Pesan error sederhana"
              },
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string",
                      "description": "Path field yang error"
                    },
                    "message": {
                      "type": "string",
                      "description": "Pesan error untuk field tersebut"
                    }
                  }
                },
                "description": "Array error detail untuk validasi"
              }
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Masukkan JWT token dengan format: Bearer <token>"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Login user",
        "description": "Endpoint untuk melakukan autentikasi user. User agent dan IP address akan diambil otomatis dari request headers. Refresh token akan disimpan di HTTP-only cookie untuk keamanan. CSRF token akan digenerate bersamaan dan dikembalikan dalam response.",
        "tags": ["Authentication"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email pengguna",
                    "example": "admin@example.com"
                  },
                  "password": {
                    "type": "string",
                    "description": "Password pengguna",
                    "example": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "tokens": {
                              "type": "object",
                              "properties": {
                                "access_token": {
                                  "type": "string",
                                  "description": "Access token untuk autentikasi",
                                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                },
                                "refresh_token": {
                                  "type": "string",
                                  "description": "Refresh token untuk memperbarui access token",
                                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                },
                                "csrf_token": {
                                  "type": "string",
                                  "description": "CSRF token untuk keamanan request POST/PUT/DELETE",
                                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                }
                              }
                            },
                            "user": {
                              "$ref": "#/components/schemas/User"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Email atau password kosong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Kredensial salah",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/refresh": {
      "post": {
        "summary": "Refresh access token",
        "description": "Endpoint untuk memperbarui access token menggunakan refresh token dari HTTP-only cookie. Refresh token akan otomatis diambil dari cookie 'refresh_token' yang diset saat login.",
        "tags": ["Authentication"],
        "security": [],
        "parameters": [
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": true,
            "description": "Refresh token dari HTTP-only cookie yang diset saat login",
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token berhasil diperbarui",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "tokens": {
                              "type": "object",
                              "properties": {
                                "access_token": {
                                  "type": "string",
                                  "description": "Access token baru",
                                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                }
                              }
                            },
                            "user": {
                              "$ref": "#/components/schemas/User"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Refresh token tidak ditemukan di cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Refresh token tidak valid atau expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}